move helper code to new setup.rs file
