+++ benches/dapol.rs
@@ -1,21 +1,13 @@
-use bulletproofs::PedersenGens;
+mod setup;
+
 use criterion::{criterion_group, criterion_main};
 use criterion::{BatchSize, BenchmarkId, Criterion, SamplingMode};
-use curve25519_dalek_ng::ristretto::RistrettoPoint;
-use curve25519_dalek_ng::scalar::Scalar;
 use iai_callgrind::{black_box, library_benchmark, library_benchmark_group, main};
 use primitive_types::H256;
-use rand::distributions::Uniform;
-use rand::Rng;
-use serde::Serialize;
 
-use core::fmt::Debug;
-
-use dapol::binary_tree::{
-    BinaryTree, Coordinate, InputLeafNode, Mergeable, Node, PathSiblings, TreeBuilder,
-};
+use dapol::binary_tree::{BinaryTree, Node};
 use dapol::node_content::FullNodeContent;
-use dapol::{AggregationFactor, Hasher, Height, InclusionProof};
+use dapol::{Height, InclusionProof};
 
 // CONSTANTS
 // ================================================================================================
@@ -48,38 +40,7 @@ const NUM_USERS: [usize; 23] = [
     250_000_000,
 ];
 
-// STRUCTS
-// ================================================================================================
-
-#[derive(Clone, Debug, PartialEq, Serialize)]
-struct BenchTestContent {
-    pub value: u32,
-    pub hash: H256,
-}
-
-impl Mergeable for BenchTestContent {
-    fn merge(left_sibling: &Self, right_sibling: &Self) -> Self {
-        // C(parent) = C(L) + C(R)
-        let parent_value = left_sibling.value + right_sibling.value;
-
-        // H(parent) = Hash(C(L) | C(R) | H(L) | H(R))
-        let parent_hash = {
-            let mut hasher = Hasher::new();
-            hasher.update(&left_sibling.value.to_le_bytes());
-            hasher.update(&right_sibling.value.to_le_bytes());
-            hasher.update(left_sibling.hash.as_bytes());
-            hasher.update(right_sibling.hash.as_bytes());
-            hasher.finalize()
-        };
-
-        BenchTestContent {
-            value: parent_value,
-            hash: parent_hash,
-        }
-    }
-}
-
-// BENCHMARKS
+// BENCHMARKS: CRITERION
 // ================================================================================================
 
 fn bench_build_tree(c: &mut Criterion) {
@@ -92,8 +53,8 @@ fn bench_build_tree(c: &mut Criterion) {
     group.bench_function(BenchmarkId::new("height_4", 8), |bench| {
         bench.iter(|| {
             let tree_height = Height::from(4);
-            let leaf_nodes = get_input_leaf_nodes(8, &tree_height);
-            build_tree(tree_height, leaf_nodes, get_padding_node_content());
+            let leaf_nodes = setup::get_input_leaf_nodes(8, &tree_height);
+            setup::build_tree(tree_height, leaf_nodes, setup::get_padding_node_content());
             ()
         })
     });
@@ -102,8 +63,8 @@ fn bench_build_tree(c: &mut Criterion) {
     group.bench_function(BenchmarkId::new("height_8", 128), |bench| {
         bench.iter(|| {
             let tree_height = Height::from(8);
-            let leaf_nodes = get_input_leaf_nodes(128, &tree_height);
-            build_tree(tree_height, leaf_nodes, get_padding_node_content());
+            let leaf_nodes = setup::get_input_leaf_nodes(128, &tree_height);
+            setup::build_tree(tree_height, leaf_nodes, setup::get_padding_node_content());
             ()
         })
     });
@@ -113,8 +74,8 @@ fn bench_build_tree(c: &mut Criterion) {
         group.bench_function(BenchmarkId::new("height_16", l), |bench| {
             bench.iter(|| {
                 let tree_height = Height::from(TREE_HEIGHTS[0]);
-                let leaf_nodes = get_input_leaf_nodes(*l, &tree_height);
-                build_tree(tree_height, leaf_nodes, get_padding_node_content());
+                let leaf_nodes = setup::get_input_leaf_nodes(*l, &tree_height);
+                setup::build_tree(tree_height, leaf_nodes, setup::get_padding_node_content());
                 ()
             })
         });
@@ -125,8 +86,8 @@ fn bench_build_tree(c: &mut Criterion) {
         group.bench_function(BenchmarkId::new("height_32", l), |bench| {
             bench.iter(|| {
                 let tree_height = Height::from(TREE_HEIGHTS[1]);
-                let leaf_nodes = get_input_leaf_nodes(*l, &tree_height);
-                build_tree(tree_height, leaf_nodes, get_padding_node_content());
+                let leaf_nodes = setup::get_input_leaf_nodes(*l, &tree_height);
+                setup::build_tree(tree_height, leaf_nodes, setup::get_padding_node_content());
                 ()
             })
         });
@@ -137,9 +98,9 @@ fn bench_build_tree(c: &mut Criterion) {
         group.bench_function(BenchmarkId::new("height_64", l), |bench| {
             bench.iter(|| {
                 let tree_height = Height::from(TREE_HEIGHTS[2]);
-                let leaf_nodes = get_input_leaf_nodes(*l, &tree_height);
+                let leaf_nodes = setup::get_input_leaf_nodes(*l, &tree_height);
 
-                build_tree(tree_height, leaf_nodes, get_padding_node_content());
+                setup::build_tree(tree_height, leaf_nodes, setup::get_padding_node_content());
                 ()
             })
         });
@@ -154,14 +115,14 @@ fn bench_generate_proof(c: &mut Criterion) {
 
     for h in TREE_HEIGHTS.into_iter() {
         let height = Height::from(h);
-        let leaf_nodes = get_full_node_contents();
+        let leaf_nodes = setup::get_full_node_contents();
 
-        let tree = build_tree(height, leaf_nodes.1, get_full_padding_node_content());
+        let tree = setup::build_tree(height, leaf_nodes.1, setup::get_full_padding_node_content());
         let leaf_node = leaf_nodes.0;
 
         group.bench_function(BenchmarkId::new("generate_proof", h), |bench| {
             bench.iter(|| {
-                generate_proof(&tree, &leaf_node);
+                setup::generate_proof(&tree, &leaf_node);
             });
         });
     }
@@ -175,16 +136,16 @@ fn bench_verify_proof(c: &mut Criterion) {
 
     for h in TREE_HEIGHTS.into_iter() {
         let height = Height::from(h);
-        let leaf_nodes = get_full_node_contents();
+        let leaf_nodes = setup::get_full_node_contents();
 
-        let tree = build_tree(height, leaf_nodes.1, get_full_padding_node_content());
+        let tree = setup::build_tree(height, leaf_nodes.1, setup::get_full_padding_node_content());
         let leaf_node = leaf_nodes.0;
 
         let root_hash = leaf_nodes.3;
 
         group.bench_function(BenchmarkId::new("verify_proof", h), |bench| {
             bench.iter_batched(
-                || generate_proof(&tree, &leaf_node),
+                || setup::generate_proof(&tree, &leaf_node),
                 |proof| proof.verify(root_hash),
                 BatchSize::SmallInput,
             );
@@ -194,41 +155,52 @@ fn bench_verify_proof(c: &mut Criterion) {
     group.finish();
 }
 
+// BENCHMARKS: IAI
+// ================================================================================================
+
 fn setup_generate(
     tree_height: Height,
 ) -> (BinaryTree<FullNodeContent>, Node<FullNodeContent>, H256) {
-    let leaf_nodes = get_full_node_contents();
-    let tree = build_tree(tree_height, leaf_nodes.1, get_full_padding_node_content());
+    let leaf_nodes = setup::get_full_node_contents();
+    let tree = setup::build_tree(
+        tree_height,
+        leaf_nodes.1,
+        setup::get_full_padding_node_content(),
+    );
 
     (tree, leaf_nodes.0, leaf_nodes.3)
 }
 
 fn setup_verify(tree_height: Height) -> (InclusionProof, H256) {
-    let leaf_nodes = get_full_node_contents();
-    let tree = build_tree(tree_height, leaf_nodes.1, get_full_padding_node_content());
+    let leaf_nodes = setup::get_full_node_contents();
+    let tree = setup::build_tree(
+        tree_height,
+        leaf_nodes.1,
+        setup::get_full_padding_node_content(),
+    );
 
-    (generate_proof(&tree, &leaf_nodes.0), leaf_nodes.3)
+    (setup::generate_proof(&tree, &leaf_nodes.0), leaf_nodes.3)
 }
 
 #[library_benchmark]
 fn bench_build_height4() -> () {
     let tree_height = Height::from(4);
-    let leaf_nodes = get_input_leaf_nodes(8, &tree_height);
-    black_box(build_tree(
+    let leaf_nodes = setup::get_input_leaf_nodes(8, &tree_height);
+    black_box(setup::build_tree(
         tree_height,
         leaf_nodes,
-        get_padding_node_content(),
+        setup::get_padding_node_content(),
     ));
 }
 
 #[library_benchmark]
 fn bench_build_height8() -> () {
     let tree_height = Height::from(8);
-    let leaf_nodes = get_input_leaf_nodes(128, &tree_height);
-    black_box(build_tree(
+    let leaf_nodes = setup::get_input_leaf_nodes(128, &tree_height);
+    black_box(setup::build_tree(
         tree_height,
         leaf_nodes,
-        get_padding_node_content(),
+        setup::get_padding_node_content(),
     ));
 }
 
@@ -236,11 +208,11 @@ fn bench_build_height8() -> () {
 fn bench_build_height16() -> () {
     for l in NUM_USERS[0..2].into_iter() {
         let tree_height = Height::from(TREE_HEIGHTS[0]);
-        let leaf_nodes = get_input_leaf_nodes(*l, &tree_height);
-        black_box(build_tree(
+        let leaf_nodes = setup::get_input_leaf_nodes(*l, &tree_height);
+        black_box(setup::build_tree(
             tree_height,
             leaf_nodes,
-            get_padding_node_content(),
+            setup::get_padding_node_content(),
         ));
     }
 }
@@ -249,11 +221,11 @@ fn bench_build_height16() -> () {
 fn bench_build_height32() -> () {
     for l in NUM_USERS[0..16].into_iter() {
         let tree_height = Height::from(TREE_HEIGHTS[1]);
-        let leaf_nodes = get_input_leaf_nodes(*l, &tree_height);
-        black_box(build_tree(
+        let leaf_nodes = setup::get_input_leaf_nodes(*l, &tree_height);
+        black_box(setup::build_tree(
             tree_height,
             leaf_nodes,
-            get_padding_node_content(),
+            setup::get_padding_node_content(),
         ));
     }
 }
@@ -262,18 +234,18 @@ fn bench_build_height32() -> () {
 fn bench_build_height64() -> () {
     for l in NUM_USERS[0..16].into_iter() {
         let tree_height = Height::from(TREE_HEIGHTS[2]);
-        let leaf_nodes = get_input_leaf_nodes(*l, &tree_height);
-        black_box(build_tree(
+        let leaf_nodes = setup::get_input_leaf_nodes(*l, &tree_height);
+        black_box(setup::build_tree(
             tree_height,
             leaf_nodes,
-            get_padding_node_content(),
+            setup::get_padding_node_content(),
         ));
     }
 }
 
 #[library_benchmark]
 fn bench_generate_height4() -> InclusionProof {
-    black_box(generate_proof(
+    black_box(setup::generate_proof(
         &setup_generate(Height::from(4)).0,
         &setup_generate(Height::from(4)).1,
     ))
@@ -281,7 +253,7 @@ fn bench_generate_height4() -> InclusionProof {
 
 #[library_benchmark]
 fn bench_generate_height8() -> InclusionProof {
-    black_box(generate_proof(
+    black_box(setup::generate_proof(
         &setup_generate(Height::from(8)).0,
         &setup_generate(Height::from(8)).1,
     ))
@@ -289,7 +261,7 @@ fn bench_generate_height8() -> InclusionProof {
 
 #[library_benchmark]
 fn bench_generate_height16() -> InclusionProof {
-    black_box(generate_proof(
+    black_box(setup::generate_proof(
         &setup_generate(Height::from(16)).0,
         &setup_generate(Height::from(16)).1,
     ))
@@ -297,7 +269,7 @@ fn bench_generate_height16() -> InclusionProof {
 
 #[library_benchmark]
 fn bench_generate_height32() -> InclusionProof {
-    black_box(generate_proof(
+    black_box(setup::generate_proof(
         &setup_generate(Height::from(32)).0,
         &setup_generate(Height::from(32)).1,
     ))
@@ -305,7 +277,7 @@ fn bench_generate_height32() -> InclusionProof {
 
 #[library_benchmark]
 fn bench_generate_height64() -> InclusionProof {
-    black_box(generate_proof(
+    black_box(setup::generate_proof(
         &setup_generate(Height::from(64)).0,
         &setup_generate(Height::from(64)).1,
     ))
@@ -349,232 +321,6 @@ fn bench_verify_height64() -> () {
     let root_hash = setup_verify(Height::from(64)).1;
 
     black_box(dapol::InclusionProof::verify(&proof, root_hash).expect("Unable to verify proof"))
-}
-
-// HELPER FUNCTIONS
-// ================================================================================================
-
-fn build_tree<C, F>(
-    height: Height,
-    leaf_nodes: Vec<Node<C>>,
-    new_padding_node_content: F,
-) -> BinaryTree<C>
-where
-    C: Clone + Debug + Mergeable + Serialize + Send + Sync + 'static,
-    F: Fn(&Coordinate) -> C + Send + Sync + 'static,
-{
-    let mut input_leaf_nodes: Vec<InputLeafNode<C>> = Vec::new();
-
-    leaf_nodes.into_iter().for_each(|n| {
-        input_leaf_nodes.push(InputLeafNode {
-            content: n.content.clone(),
-            x_coord: n.coord.x,
-        })
-    });
-
-    let builder = TreeBuilder::<C>::new()
-        .with_height(height)
-        .with_leaf_nodes(input_leaf_nodes);
-
-    let tree = builder
-        .build_using_multi_threaded_algorithm(new_padding_node_content)
-        .expect("Unable to build tree");
-
-    tree
-}
-
-fn generate_proof(
-    tree: &BinaryTree<FullNodeContent>,
-    leaf_node: &Node<FullNodeContent>,
-) -> InclusionProof {
-    let aggregation_factor = AggregationFactor::Divisor(2u8);
-    let upper_bound_bit_length = 64u8;
-
-    let path_siblings = PathSiblings::build_using_multi_threaded_algorithm(
-        tree,
-        leaf_node,
-        get_full_padding_node_content(),
-    )
-    .expect("Unable to generate path siblings");
-
-    let proof = InclusionProof::generate(
-        leaf_node.clone(),
-        path_siblings,
-        aggregation_factor,
-        upper_bound_bit_length,
-    )
-    .expect("Unable to generate proof");
-
-    proof
-}
-
-fn get_input_leaf_nodes(num_leaves: usize, height: &Height) -> Vec<Node<BenchTestContent>> {
-    let max_bottom_layer_nodes = 2usize.pow(height.as_u32() - 1);
-
-    assert!(
-        num_leaves <= max_bottom_layer_nodes,
-        "Number of leaves exceeds maximum bottom layer nodes"
-    );
-
-    let mut leaf_nodes: Vec<Node<BenchTestContent>> = Vec::new();
-
-    for i in 0..num_leaves {
-        leaf_nodes.push(
-            InputLeafNode::<BenchTestContent> {
-                x_coord: i as u64,
-                content: BenchTestContent {
-                    hash: H256::random(),
-                    value: i as u32,
-                },
-            }
-            .into_node(),
-        );
-    }
-
-    leaf_nodes
-}
-
-fn get_full_node_contents(// height: &Height,
-) -> (
-    Node<FullNodeContent>,
-    Vec<Node<FullNodeContent>>,
-    RistrettoPoint,
-    H256,
-) {
-    let mut rng = rand::thread_rng();
-    let liability_range = Uniform::new(1, u64::MAX);
-
-    let liabilities = [
-        rng.sample(liability_range),
-        rng.sample(liability_range),
-        rng.sample(liability_range),
-        rng.sample(liability_range),
-    ];
-
-    let blinding_factors = [
-        Scalar::from_bytes_mod_order(*b"90998600161833439099840024221618"),
-        Scalar::from_bytes_mod_order(*b"34334644060024221618334357559098"),
-        Scalar::from_bytes_mod_order(*b"16183433909906002422161834335755"),
-        Scalar::from_bytes_mod_order(*b"46442422181134335755929806001618"),
-    ];
-
-    let commitments = [
-        PedersenGens::default().commit(Scalar::from(liabilities[0]), blinding_factors[0]),
-        PedersenGens::default().commit(Scalar::from(liabilities[1]), blinding_factors[1]),
-        PedersenGens::default().commit(Scalar::from(liabilities[2]), blinding_factors[2]),
-        PedersenGens::default().commit(Scalar::from(liabilities[3]), blinding_factors[3]),
-    ];
-
-    let bytes = [
-        b"leafleafleafleafleafleafleafleaf",
-        b"sibling1sibling1sibling1sibling1",
-        b"sibling2sibling2sibling2sibling2",
-        b"sibling3sibling3sibling3sibling3",
-    ];
-
-    let mut hasher = Hasher::new();
-    hasher.update(bytes[0]);
-    let hash = hasher.finalize();
-
-    let leaf = Node {
-        coord: Coordinate { x: 2u64, y: 0u8 },
-        content: FullNodeContent::new(liabilities[0], blinding_factors[0], commitments[0], hash),
-    };
-
-    let mut hasher = Hasher::new();
-    hasher.update(bytes[1]);
-    let hash = hasher.finalize();
-
-    let sibling1 = Node {
-        coord: Coordinate { x: 3u64, y: 0u8 },
-        content: FullNodeContent::new(liabilities[1], blinding_factors[1], commitments[1], hash),
-    };
-
-    let (parent_commitment, parent_hash) = get_parent_node(
-        leaf.content.commitment,
-        sibling1.content.commitment,
-        leaf.content.hash,
-        sibling1.content.hash,
-    );
-
-    let mut hasher = Hasher::new();
-    hasher.update(bytes[2]);
-    let hash = hasher.finalize();
-
-    let sibling2 = Node {
-        coord: Coordinate { x: 0u64, y: 1u8 },
-        content: FullNodeContent::new(liabilities[2], blinding_factors[2], commitments[2], hash),
-    };
-
-    let (parent_commitment, parent_hash) = get_parent_node(
-        sibling2.content.commitment,
-        parent_commitment,
-        sibling2.content.hash,
-        parent_hash,
-    );
-
-    let mut hasher = Hasher::new();
-    hasher.update(bytes[3]);
-    let hash = hasher.finalize();
-
-    let sibling3 = Node {
-        coord: Coordinate { x: 1u64, y: 2u8 },
-        content: FullNodeContent::new(liabilities[3], blinding_factors[3], commitments[3], hash),
-    };
-
-    let (root_commitment, root_hash) = get_parent_node(
-        sibling3.content.commitment,
-        parent_commitment,
-        sibling3.content.hash,
-        parent_hash,
-    );
-
-    let nodes = vec![leaf.clone(), sibling1, sibling2, sibling3];
-
-    (leaf, nodes, root_commitment, root_hash)
-}
-
-fn get_parent_node(
-    left_commitment: RistrettoPoint,
-    right_commitment: RistrettoPoint,
-    left_hash: H256,
-    right_hash: H256,
-) -> (RistrettoPoint, H256) {
-    let parent_commitment: RistrettoPoint = left_commitment + right_commitment;
-
-    let mut hasher = Hasher::new();
-
-    let parent_hash: H256 = {
-        hasher.update(left_commitment.compress().as_bytes());
-        hasher.update(right_commitment.compress().as_bytes());
-        hasher.update(left_hash.as_bytes());
-        hasher.update(right_hash.as_bytes());
-        hasher.finalize()
-    };
-
-    (parent_commitment, parent_hash)
-}
-
-fn get_padding_node_content() -> impl Fn(&Coordinate) -> BenchTestContent {
-    |_coord: &Coordinate| -> BenchTestContent {
-        BenchTestContent {
-            value: 0,
-            hash: H256::default(),
-        }
-    }
-}
-
-fn get_full_padding_node_content() -> impl Fn(&Coordinate) -> FullNodeContent {
-    |_coord: &Coordinate| -> FullNodeContent {
-        let liability = 27u64;
-        let blinding_factor = Scalar::from_bytes_mod_order(*b"11112222333344445555666677778888");
-        let commitment = PedersenGens::default().commit(Scalar::from(liability), blinding_factor);
-        let mut hasher = Hasher::new();
-        hasher.update("leaf".as_bytes());
-        let hash = hasher.finalize();
-
-        FullNodeContent::new(liability, blinding_factor, commitment, hash)
-    }
 }
 
 criterion_group!(
